+++ daemon/src/rollover/maker.rs
@@ -1,7 +1,5 @@
-use crate::rollover::protocol;
 use crate::rollover::protocol::*;
 use anyhow::Context;
-use anyhow::Result;
 use async_trait::async_trait;
 use asynchronous_codec::Framed;
 use asynchronous_codec::JsonCodec;
@@ -10,15 +8,12 @@ use futures::SinkExt;
 use futures::StreamExt;
 use libp2p_core::PeerId;
 use maia_core::secp256k1_zkp::XOnlyPublicKey;
-use model::BaseDlcParams;
 use model::Dlc;
 use model::ExecuteOnCfd;
-use model::FundingRate;
 use model::OrderId;
 use model::Position;
 use model::Role;
 use model::RolloverVersion;
-use model::TxFeeRate;
 use std::collections::HashMap;
 use tokio_extras::FutureExt;
 use tokio_extras::Tasks;
@@ -26,44 +21,47 @@ use xtra_libp2p::NewInboundSubstream;
 use xtra_libp2p::Substream;
 use xtra_productivity::xtra_productivity;
 
-type ListenerConnection = (
-    Framed<Substream, JsonCodec<ListenerMessage, DialerMessage>>,
-    PeerId,
-    BaseDlcParams,
-);
-
 /// Permanent actor to handle incoming substreams for the `/itchysats/rollover/1.0.0`
 /// protocol.
 ///
 /// There is only one instance of this actor for all connections, meaning we must always spawn a
 /// task whenever we interact with a substream to not block the execution of other connections.
-pub struct Actor<E, O> {
+pub struct Actor<E, O, R> {
     protocol_tasks: HashMap<OrderId, Tasks>,
     oracle_pk: XOnlyPublicKey,
     oracle: O,
     n_payouts: usize,
-    pending_protocols: HashMap<OrderId, ListenerConnection>,
     executor: E,
+    rates: R,
+    is_accepting_rollovers: bool,
 }
 
-impl<E, O> Actor<E, O> {
-    pub fn new(executor: E, oracle_pk: XOnlyPublicKey, oracle: O, n_payouts: usize) -> Self {
+impl<E, O, R> Actor<E, O, R> {
+    pub fn new(
+        executor: E,
+        oracle_pk: XOnlyPublicKey,
+        oracle: O,
+        rates: R,
+        n_payouts: usize,
+    ) -> Self {
         Self {
             protocol_tasks: HashMap::default(),
             oracle_pk,
             oracle,
             n_payouts,
-            pending_protocols: HashMap::default(),
             executor,
+            rates,
+            is_accepting_rollovers: true,
         }
     }
 }
 
 #[async_trait]
-impl<E, O> xtra::Actor for Actor<E, O>
+impl<E, O, R> xtra::Actor for Actor<E, O, R>
 where
     E: Send + Sync + 'static,
     O: Send + Sync + 'static,
+    R: Send + Sync + 'static,
 {
     type Stop = ();
 
@@ -71,10 +69,23 @@ where
 }
 
 #[xtra_productivity]
-impl<E, O> Actor<E, O>
+impl<E, O, R> Actor<E, O, R>
 where
     E: ExecuteOnCfd + Clone + Send + Sync + 'static,
     O: GetAnnouncements + Clone + Send + Sync + 'static,
+    R: GetRates + Clone + Send + Sync + 'static,
+{
+    async fn handle(&mut self, msg: UpdateConfiguration) {
+        self.is_accepting_rollovers = msg.is_accepting_rollovers;
+    }
+}
+
+#[xtra_productivity]
+impl<E, O, R> Actor<E, O, R>
+where
+    E: ExecuteOnCfd + Clone + Send + Sync + 'static,
+    O: GetAnnouncements + Clone + Send + Sync + 'static,
+    R: GetRates + Clone + Send + Sync + 'static,
 {
     async fn handle(&mut self, msg: NewInboundSubstream, ctx: &mut xtra::Context<Self>) {
         let NewInboundSubstream { peer, stream } = msg;
@@ -112,7 +123,7 @@ where
     async fn handle(&mut self, msg: ProposeReceived) {
         let ProposeReceived {
             propose,
-            framed,
+            mut framed,
             peer,
         } = msg;
         let order_id = propose.order_id;
@@ -138,43 +149,54 @@ where
             }
         };
 
-        // In case we fail to accept/reject some proposals might never get cleaned up. Given that
-        // the taker will retry this should not do much harm because we will replace the proposal
-        // with a new one. This is acceptable for now.
-        self.pending_protocols
-            .insert(order_id, (framed, peer, base_dlc_params));
-    }
+        if !self.is_accepting_rollovers {
+            emit_rejected(order_id, &self.executor).await;
 
-    async fn handle(&mut self, msg: Accept) -> Result<()> {
-        let Accept {
-            order_id,
-            tx_fee_rate,
-            long_funding_rate,
-            short_funding_rate,
-        } = msg;
+            let mut tasks = Tasks::default();
+            tasks.add_fallible(
+                async move {
+                    framed
+                        .send(ListenerMessage::Decision(Decision::Reject(Reject {
+                            order_id,
+                        })))
+                        .await
+                },
+                move |e| async move {
+                    tracing::debug!(%order_id, "Failed to send reject rollover to the taker: {e:#}")
+                },
+            );
+            self.protocol_tasks.insert(order_id, tasks);
+
+            return;
+        }
 
         fn next_rollover_span() -> tracing::Span {
             tracing::debug_span!("next rollover message")
         }
-
-        let (mut framed, _, base_dlc_params) =
-            self.pending_protocols.remove(&order_id).with_context(|| {
-                format!("No active protocol for {order_id} when accepting rollover")
-            })?;
 
         let mut tasks = Tasks::default();
         tasks.add_fallible(
             {
                 let executor = self.executor.clone();
                 let oracle = self.oracle.clone();
+                let rates =  self.rates.clone();
                 let oracle_pk = self.oracle_pk;
                 let n_payouts = self.n_payouts;
                 async move {
+                    let Rates {
+                        funding_rate_long,
+                        funding_rate_short,
+                        tx_fee_rate,
+                    } = rates
+                        .get_rates()
+                        .await
+                        .context("Failed to get rates")?;
+
                     let (rollover_params, dlc, position, oracle_event_id, funding_rate) = executor
                         .execute(order_id, |cfd| {
                             let funding_rate = match cfd.position() {
-                                Position::Long => long_funding_rate,
-                                Position::Short => short_funding_rate,
+                                Position::Long => funding_rate_long,
+                                Position::Short => funding_rate_short,
                             };
 
                             let (event, params, dlc, position, oracle_event_id) = cfd
@@ -350,35 +372,19 @@ where
             },
         );
         self.protocol_tasks.insert(order_id, tasks);
-
-        Ok(())
     }
+}
 
-    async fn handle(&mut self, msg: Reject) -> Result<()> {
-        let Reject { order_id } = msg;
+#[derive(Clone, Copy)]
+pub struct UpdateConfiguration {
+    is_accepting_rollovers: bool,
+}
 
-        let (mut framed, ..) = self.pending_protocols.remove(&order_id).with_context(|| {
-            format!("No active protocol for {order_id} when rejecting rollover")
-        })?;
-
-        emit_rejected(order_id, &self.executor).await;
-
-        let mut tasks = Tasks::default();
-        tasks.add_fallible(
-            async move {
-                framed
-                    .send(ListenerMessage::Decision(Decision::Reject(
-                        protocol::Reject { order_id },
-                    )))
-                    .await
-            },
-            move |e| async move {
-                tracing::debug!(%order_id, "Failed to send reject rollover to the taker: {e:#}")
-            },
-        );
-        self.protocol_tasks.insert(order_id, tasks);
-
-        Ok(())
+impl UpdateConfiguration {
+    pub fn new(is_accepting_rollovers: bool) -> Self {
+        Self {
+            is_accepting_rollovers,
+        }
     }
 }
 
@@ -386,19 +392,4 @@ struct ProposeReceived {
     propose: Propose,
     framed: Framed<Substream, JsonCodec<ListenerMessage, DialerMessage>>,
     peer: PeerId,
-}
-
-/// Upon accepting Rollover maker sends the current estimated transaction fee and
-/// funding rate
-#[derive(Clone, Copy, Debug)]
-pub struct Accept {
-    pub order_id: OrderId,
-    pub tx_fee_rate: TxFeeRate,
-    pub long_funding_rate: FundingRate,
-    pub short_funding_rate: FundingRate,
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct Reject {
-    pub order_id: OrderId,
 }